requirements.txt
Django==3.2.4
django-allauth==0.45.0
django-core==1.4.1
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==21.1
django-heroku==0.3.1
django-rest-auth==0.9.5
python-decouple==3.6
djangorestframework==3.13.1